// packages/studio-app/src/studioDataSources/movies/types.ts
export interface MoviesConnectionParams {
  apiKey: string;
}

export interface MoviesQuery {
  genre: string | null;
}

export interface Movie {
  id: number;
  title: string;
  year: string;
  runtime: string;
  genres: string[];
  director: string;
  actors: string;
  plot: string;
  posterUrl: string;
}